refactor: merge TestContextBase and TestContext
