+++ MIGRATION.md
@@ -48,3 +48,6 @@ The `Fake` prefix has been replaced with `Bunit` in many test doubles. For examp
 
 ### Renamed `AddTestAuthorization` to `AddAuthorization`
 The `AddTestAuthorization` method on `TestContext` has been renamed to `AddAuthorization`. If you used `AddTestAuthorization`, you should replace it with `AddAuthorization`.
+
+## Merged `TestContext` and `TestContextBase`
+The `TestContext` and `TestContextBase` classes have been merged into a single `TestContext` class. If you used either of these classes, you should replace them with `TestContext`.
