+++ src/bunit/Extensions/TestContextBaseRenderExtensions.cs
@@ -3,18 +3,18 @@ using Bunit.Rendering;
 namespace Bunit.Extensions;
 
 /// <summary>
-/// Extensions methods for <see cref="TestContextBase"/> types.
+/// Extensions methods for <see cref="TestContext"/> types.
 /// </summary>
-public static class TestContextBaseRenderExtensions
+public static class TestContextRenderExtensions
 {
 	/// <summary>
-	/// Renders a component, declared in the <paramref name="renderFragment"/>, inside the <see cref="TestContextBase.RenderTree"/>.
+	/// Renders a component, declared in the <paramref name="renderFragment"/>, inside the <see cref="TestContext.RenderTree"/>.
 	/// </summary>
 	/// <typeparam name="TComponent">The type of component to render.</typeparam>
 	/// <param name="testContext">Test context to use to render with.</param>
 	/// <param name="renderFragment">The <see cref="RenderInsideRenderTree"/> that contains a declaration of the component.</param>
 	/// <returns>A <see cref="IRenderedComponent{TComponent}"/>.</returns>
-	public static IRenderedComponent<TComponent> RenderInsideRenderTree<TComponent>(this TestContextBase testContext, RenderFragment renderFragment)
+	public static IRenderedComponent<TComponent> RenderInsideRenderTree<TComponent>(this TestContext testContext, RenderFragment renderFragment)
 		where TComponent : IComponent
 	{
 		ArgumentNullException.ThrowIfNull(testContext);
@@ -24,12 +24,12 @@ public static class TestContextBaseRenderExtensions
 	}
 
 	/// <summary>
-	/// Renders a fragment, declared in the <paramref name="renderFragment"/>, inside the <see cref="TestContextBase.RenderTree"/>.
+	/// Renders a fragment, declared in the <paramref name="renderFragment"/>, inside the <see cref="TestContext.RenderTree"/>.
 	/// </summary>
 	/// <param name="testContext">Test context to use to render with.</param>
 	/// <param name="renderFragment">The <see cref="RenderInsideRenderTree"/> to render.</param>
 	/// <returns>A <see cref="IRenderedFragment"/>.</returns>
-	public static IRenderedFragment RenderInsideRenderTree(this TestContextBase testContext, RenderFragment renderFragment)
+	public static IRenderedFragment RenderInsideRenderTree(this TestContext testContext, RenderFragment renderFragment)
 	{
 		ArgumentNullException.ThrowIfNull(testContext);
